refactor(catalog): move federation to crate (Slice 3)
Third catalog-extraction increment. Moves the federation module
(src/catalog/federation/{mod,external/mod,federated_catalog}) into
proximadb-catalog. These were deferred in Slice 1 because they used CatalogManager
-- which Slice 2 moved into the crate, clearing the blocker.

The federation files are foundation-clean (only CatalogManager + the Catalog trait
+ intra-catalog deps; no storage/query/services/index/network up-edges, no web
stack). Path rewrites: crate::catalog::{internal,traits,federation} -> crate::
(crate::catalog::traits::Catalog -> crate::Catalog since the crate has no `traits`
shim module). Root src/catalog/mod.rs re-exports federation from the crate
(pub use proximadb_catalog::federation;) so all crate::catalog::federation::*
paths are unchanged.

iceberg_rest_service STAYS root -- it uses
proximadb_storage_common::object_store_bridge::ObjectStoreBridge, and
storage-common -> catalog is a downward edge, so catalog -> storage-common would
cycle (the same trap as segment_registry). It moves once ObjectStoreBridge is
hoisted to a shared foundation crate.

Verification: crate + root clippy --lib --bins -D warnings clean; fmt clean;
layering clean (storage up-edge ratchet unchanged at 197); 224 root catalog
tests + 434 catalog-crate tests pass.
