+++ crates/control/proximadb-catalog/src/federation/mod.rs
